fix bug invite email and api team and position

+++ src/main/java/com/ces/intern/hr/resourcing/demo/sevice/impl/TeamServiceImpl.java
@@ -22,6 +22,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -32,24 +33,19 @@ public class TeamServiceImpl implements TeamService {
     private final ResourceRepository resourceRepository;
     private final WorkspaceRepository workspaceRepository;
     private final PositionRepository positionRepository;
-    private final TeamService teamService;
-    private final PositionService positionService;
 
     @Autowired
     public TeamServiceImpl(TeamRepository teamRepository,
                            ModelMapper modelMapper,
                            ResourceRepository resourceRepository,
                            WorkspaceRepository workspaceRepository,
-                           PositionRepository positionRepository,
-                           TeamService teamService,
-                           PositionService positionService) {
+                           PositionRepository positionRepository
+                   ) {
         this.teamRepository = teamRepository;
         this.modelMapper = modelMapper;
         this.resourceRepository = resourceRepository;
         this.workspaceRepository = workspaceRepository;
         this.positionRepository = positionRepository;
-        this.teamService = teamService;
-        this.positionService = positionService;
     }
 
     @Override
@@ -60,16 +56,7 @@ public class TeamServiceImpl implements TeamService {
         return teamEntityList.stream().map(s -> modelMapper.map(s, TeamDTO.class)).collect(Collectors.toList());
     }
 
-    @Override
-    public void addResourceToTeam(Integer idTeam, Integer idResource) {
-        ResourceEntity resourceEntity = resourceRepository.findById(idResource)
-                .orElseThrow(() -> new NotFoundException(ExceptionMessage.NOT_FOUND_RECORD.getMessage()));
-        TeamEntity teamEntity = teamRepository.findById(idTeam)
-                .orElseThrow(() -> new NotFoundException(ExceptionMessage.NOT_FOUND_RECORD.getMessage()));
-        resourceEntity.getPositionEntity().setTeamEntity(teamEntity);
-        resourceRepository.save(resourceEntity);
 
-    }
 
     @Override
     public void deleteTeam(Integer idTeam) {
@@ -79,8 +66,8 @@ public class TeamServiceImpl implements TeamService {
     }
 
     @Override
-    public void renameTeam(Integer idTeam, String name) {
-        TeamEntity teamEntity = teamRepository.findById(idTeam)
+    public void renameTeam(Integer idWorkspace,Integer idTeam, String name) {
+        TeamEntity teamEntity = teamRepository.findByidWorkspaceAndIdTeam(idWorkspace,idTeam)
                 .orElseThrow(() -> new NotFoundException(ExceptionMessage.NOT_FOUND_RECORD.getMessage()));
         teamEntity.setName(name);
         teamRepository.save(teamEntity);
@@ -131,11 +118,12 @@ public class TeamServiceImpl implements TeamService {
 
     @Override
     public void created(List<TeamRequest> teamRequests, Integer idWorkspace) {
-        WorkspaceEntity workspaceEntity=workspaceRepository.findById(idWorkspace).get();
+        WorkspaceEntity workspaceEntity=workspaceRepository.findById(idWorkspace).orElse(null);
         for (TeamRequest teamRequest : teamRequests) {
             if (!teamRepository.findByNameAndidWorkspace(teamRequest.getName(), idWorkspace).isPresent()) {
                 TeamEntity teamEntity = new TeamEntity();
                 teamEntity.setName(teamRequest.getName());
+                teamEntity.setCreatedDate(new Date());
                 teamEntity.setWorkspaceEntityTeam(workspaceEntity);
                 teamRepository.save(teamEntity);
                 for (PositionRequest positionRequest : teamRequest.getPositions()) {
@@ -196,8 +184,31 @@ public class TeamServiceImpl implements TeamService {
     private void deleteTeam(List<TeamRequest> teamRequests, List<TeamEntity> teamEntities) {
         for (TeamEntity teamEntity : teamEntities) {
             if (teamRequests.stream().filter(s -> teamEntity.getName().equals(s.getName())).findAny().orElse(null) == null) {
-                teamRepository.delete(teamEntity);
-            }
+                    if (resourceRepository
+                            .countAllByWorkspaceEntityResource_IdAndTeamEntityResource_Id
+                                    (teamEntity.getWorkspaceEntityTeam().getId(), teamEntity.getId()) == 0){
+                        for (PositionEntity positionEntity : positionRepository.findAllByTeamEntity_Id(teamEntity.getId())){
+                            positionRepository.deleteById(positionEntity.getId());
+                        }
+                        teamRepository.deleteById(teamEntity.getId());
+                    }
+                    else {
+                        teamEntity.setIsArchived(true);
+                        for (PositionEntity positionEntity : positionRepository.findAllByTeamEntity_Id(teamEntity.getId())){
+                            if (resourceRepository
+                                    .countResourcesOfPosition
+                                            (positionEntity.getId(), positionEntity.getTeamEntity().getWorkspaceEntityTeam().getId()) == 0){
+                                positionRepository.deleteById(positionEntity.getId());
+                            }
+                            else {
+                                positionEntity.setIsArchived(true);
+                                positionRepository.save(positionEntity);
+                            }
+                        }
+                        teamRepository.save(teamEntity);
+                    }
+                }
+
         }
     }
 
@@ -205,34 +216,23 @@ public class TeamServiceImpl implements TeamService {
 
         for (PositionEntity positionEntity : positionEntities) {
             if (positionRequests.stream().filter(s -> positionEntity.getName().equals(s.getName())).findAny().orElse(null) == null) {
-                positionRepository.delete(positionEntity);
+                if (positionEntity != null){
+                    if (resourceRepository
+                            .countResourcesOfPosition
+                                    (positionEntity.getId(), positionEntity.getTeamEntity().getWorkspaceEntityTeam().getId()) == 0){
+                        positionRepository.delete(positionEntity);
+                    }
+                    else {
+                        positionEntity.setIsArchived(true);
+                        positionRepository.save(positionEntity);
+                    }
+                }
             }
         }
     }
 
 
-    @Override
-    public void deleteOneTeam(TeamRequest teamRequest) {
-        TeamEntity teamEntity = teamRepository.findById(teamRequest.getId()).orElse(null);
-        if (teamEntity != null){
-            if (resourceRepository
-                    .countAllByWorkspaceEntityResource_IdAndTeamEntityResource_Id
-                            (teamEntity.getWorkspaceEntityTeam().getId(), teamEntity.getId()) == 0){
-                for (PositionEntity positionEntity : teamEntity.getPositionEntities()){
-                    positionRepository.deleteById(positionEntity.getId());
-                }
-                teamRepository.deleteById(teamEntity.getId());
-            }
-            else {
-                teamEntity.setIsArchived(true);
-                for (PositionEntity positionEntity : teamEntity.getPositionEntities()){
-                    positionEntity.setIsArchived(true);
-                    positionRepository.save(positionEntity);
-                }
-                teamRepository.save(teamEntity);
-            }
-        }
-    }
+
 
     @Override
     public void deleteMultipleTeam(List<TeamRequest> teamRequests) {
